# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: TI-Toolbox job server (tit.server)
  version: "1.0.0"
  description: >-
    v1 contract for the v3 Electron GUI. Every path and schema from
    `contracts/openapi.v0.yaml` is carried unchanged (the real server's `--dump-openapi` output
    must be a superset of both; CI gate `dev/contracts_check.py`); v1 adds the job engine, catalog
    CRUD, ViewSpec viewers, jailed file access, validate/plan preflight, settings, and the
    `/ws/jobs` live-job stream. Auth is unchanged from v0: the launcher passes
    `?token=<TIT_SERVER_TOKEN>` exactly once; the server mints a random session id (never the
    token), keeps it in memory, sets it as an HttpOnly, SameSite=Strict `tit_session` cookie and
    redirects to `/`. A server restart forgets every session (401 -> back to the launcher). API
    routes accept either that cookie or `Authorization: Bearer <token>` (scripts). Unauthenticated
    requests to any route other than `/api/health` and `/auth/session` get 401.
    `POST /auth/logout` forgets the session and clears the cookie.

    Config request bodies (`config` fields below) are `dict[str, Any]`, validated server-side with
    `tit.config_io.deserialize_config`. Their real shapes are the `contracts/schema.json` `$defs`
    (`SimulationConfig`, `Montage`, `FlexConfig`, `ExConfig`, `MExConfig`, `AnalyzerConfig`,
    `PreprocessConfig`, `QSIPrepConfig`, `QSIReconConfig`, `GroupComparisonConfig`,
...
    today.

...

...

...
    difference is never reported as drift.
...
servers:
...
paths:
...
  /api/schema:
    get:
      tags: [schema]
      summary: The full generated schema.json (every config dataclass as a JSON Schema $def)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/schema/{name}:
    get:
      tags: [schema]
      summary: One resolved $def from schema.json, by config class name
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown config class name

  # --------------------------------------------------------------------------- validate / plan (v1)
...
  /api/pipelines:
    get:
      tags: [pipelines]
      summary: Saved pipelines in this project (code/ti-toolbox/pipelines/)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PipelineListEntry" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/pipelines/kinds:
    get:
      tags: [pipelines]
      summary: Node kinds and their typed ports (the canvas palette)
      description: >-
        The canvas draws its ports from this rather than from a hand-kept copy, so a port added
        on the server appears in the UI without a renderer change.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineKinds"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/pipelines/validate:
    post:
      tags: [pipelines]
      summary: Validate a pipeline graph, with a reason per problem
      description: >-
        Never 500s on a graph problem: a cycle, an incompatible port or an unbound required input
        comes back as an `issue` with a sentence the UI shows next to the node or wire. A 422 means
        the document is not shaped like a pipeline at all.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipelineRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineValidation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: the document is not a valid pipeline document
  /api/pipelines/run:
    post:
      tags: [pipelines]
      summary: Run a whole pipeline as one job group
      description: >-
        One submission. Every job's `after` is the document's edges resolved to real job ids, and
        every job carries the one `group_id` in the response -- so Jobs shows the pipeline as one
        row with its member tree and cancels as one thing. Ports whose value only exists after the
        producer ran (montage names from an optimizer, a leadfield, a simulation name whose montages
        were themselves optimized) get a small `tools` resolve job planned between producer and
        consumer.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipelineRunRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineRunResult" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: the pipeline does not validate (body carries the issues) or a config is invalid
  /api/pipelines/export:
    post:
      tags: [pipelines]
      summary: Export a pipeline as a Jupyter notebook
      description: >-
        A pure function of the document. The body is the `.ipynb` JSON as text; the pipeline itself
        rides along in `metadata.ti_toolbox.pipeline`, so a later import restores the canvas without
        parsing Python.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: format
          required: false
          schema: { type: string, enum: [ipynb], default: ipynb }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipelineRequest" }
      responses:
        "200":
          content:
            text/plain:
              schema: { type: string, description: the .ipynb document as JSON text }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: unsupported format, or an invalid document
        "501":
          description: nbformat is not installed in this environment
  /api/pipelines/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    get:
      tags: [pipelines]
      summary: Load a saved pipeline
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pipeline" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: illegal pipeline name, or the saved file is not a pipeline document
    put:
      tags: [pipelines]
      summary: Save a pipeline document
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipelineRequest" }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [name, saved]
                properties:
                  name: { type: string }
                  saved: { type: boolean }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: illegal pipeline name, or an invalid document
    delete:
      tags: [pipelines]
      summary: Delete a saved pipeline
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "204":
          description: deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ----------------------------------------------------------------------------------- tetravox (v1)
  # dev/notes/v3-embed-convergence-plan.md, decisions E1-E4. Nothing here auto-installs: every
  # write starts at a request the user made in Settings -> Viewer engine.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tit_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: missing or invalid session/token
    NotFound:
      description: no such resource
...
  schemas:
...
    JobKind:
      type: string
      enum:
        [
          pre,
          sim,
          flex,
          flex_adaptive,
          flex_pareto,
          ex,
          mex,
          leadfield,
          analyzer,
          stats,
          source,
          blender,
          nifti_average,
          nilearn,
          project_init,
          tools,
          report,
        ]
    JobState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled, skipped, lost]
    JobProgress:
      type: object
      required: [stage, i, n, pct]
      properties:
        stage: { type: string }
        i: { type: integer }
        n: { type: integer }
        pct: { type: number }
    WaitingOn:
      type: object
      required: [key, job_id]
      properties:
        key: { type: string }
        job_id: { type: string }
    JobError:
      type: object
      required: [type, message, last_lines]
      properties:
        type:
          type: string
          enum:
            [
              preflight,
              lock_wait,
              budget_wait,
              runner_failed,
              oom_suspected,
              cancelled,
              skipped,
              lost,
              docker_unavailable,
              kind_error,
            ]
          description: >-
            The nine-value taxonomy `pages/jobs/format.ts::ERROR_LABEL` labels, plus `kind_error`
            (`tit/jobs/manager.py:628`, an unknown/disallowed job kind rejected before a runner ever
            starts). `pages/jobs/format.ts::errorLabel` title-cases any other value it does not
            recognise, so a not-yet-enumerated runner-emitted type degrades gracefully rather than
            crashing the Jobs page.
        message: { type: string }
        last_lines:
          type: array
          items: { type: string }
    Artifact:
      type: object
      required: [path, kind]
      properties:
        path: { type: string }
        kind: { type: string }
        label: { type: string }
    JobStatus:
      type: object
      required: [id, kind, state, subject_ids, created_at, artifacts]
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/JobKind" }
        state: { $ref: "#/components/schemas/JobState" }
        subject_ids:
          type: array
          items: { type: string }
        group_id: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        progress:
          oneOf:
            - $ref: "#/components/schemas/JobProgress"
            - type: "null"
        liveness:
          type: [string, "null"]
          enum: [active, stalled, null]
        waiting_on:
          type: array
          items: { $ref: "#/components/schemas/WaitingOn" }
        exit_code: { type: [integer, "null"] }
        error:
          oneOf:
            - $ref: "#/components/schemas/JobError"
            - type: "null"
        artifacts:
          type: array
          items: { $ref: "#/components/schemas/Artifact" }
        cpu_percent: { type: [number, "null"] }
        rss: { type: [integer, "null"] }
        log_path:
          type: [string, "null"]
          description: >-
            Container path to this job's raw stdout/stderr log (`code/ti-toolbox/jobs/<id>/stdout.log`),
            the same file `GET /api/jobs/{id}/log` serves. Present so `pages/jobs/JobDetailDrawer.tsx`
            can show/copy it without reconstructing the path convention client-side (ra_13 finding 6).
...
    PipelineNode:
      type: object
      description: One node = exactly one existing job kind carrying exactly the config that page builds.
      required: [id, kind]
      properties:
        id: { type: string, description: canvas-scoped node id, unique within the document }
        kind:
          type: string
          enum: [pre, leadfield, flex, ex, mex, sim, analyzer, source, stats]
        label: { type: string, description: user-visible name; defaults to the kind }
        config:
          $ref: "#/components/schemas/PipelineConfig"
        position:
          type: object
          description: canvas coordinates; carried through untouched by the server
          properties:
            x: { type: number }
            y: { type: number }
    PipelineEdge:
      type: object
      description: >-
        A typed wire. `port` is one named output of `from` bound into the same-named input of `to`
        -- which is what makes "Optimizer -> Simulator" mean "the simulator's montage list is the
        optimizer's result" rather than a drawn line with no semantics.
      required: [from, to, port]
      properties:
        from: { type: string }
        to: { type: string }
        port:
          type: string
          enum: [subjects, montages, simulation, roi, leadfield]
    Pipeline:
      type: object
      description: The canvas document; mirrors contracts/pipeline.schema.json exactly.
      required: [version, nodes]
      properties:
        version: { type: integer, enum: [1] }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/PipelineNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/PipelineEdge" }
    PipelineRequest:
      type: object
      description: >-
        A pipeline document, either sent bare or wrapped as `{pipeline: <document>}` (the wrapped
        form is what `run` uses, since it carries submission options alongside).
      properties:
        pipeline: { $ref: "#/components/schemas/Pipeline" }
        version: { type: integer, enum: [1] }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/PipelineNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/PipelineEdge" }
    PipelineRunRequest:
      type: object
      required: [pipeline]
      properties:
        pipeline: { $ref: "#/components/schemas/Pipeline" }
        parallel_subjects:
          type: integer
          minimum: 1
          default: 1
          description: >-
            The same per-group concurrency cap JobGroupRequest carries; enforced by the scheduler.
        tags:
          type: array
          items: { type: string }
        overwrite:
          type: boolean
          description: Replace existing output instead of skipping it, for every job in the pipeline.
    PipelineIssue:
      type: object
      required: [level, message]
      properties:
        level: { type: string, enum: [error, warning] }
        message: { type: string, description: one sentence the UI shows verbatim }
        node_id: { type: string }
        edge: { $ref: "#/components/schemas/PipelineEdge" }
    PipelineJobPreview:
      type: object
      description: >-
        One job `run` would submit -- the receipt, computed without submitting anything. `label`
        and `after` are the plan-scoped labels the server resolves to real job ids at submission.
      required: [label, kind, subject_ids, after]
      properties:
        label: { type: string }
        kind: { type: string }
        subject_ids:
          type: array
          items: { type: string }
        after:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    PipelineValidation:
      type: object
      required: [ok, issues, order]
      properties:
        ok: { type: boolean, description: false iff any issue has level "error" }
        issues:
          type: array
          items: { $ref: "#/components/schemas/PipelineIssue" }
        order:
          type: array
          description: node ids in dependency-first order; empty when the graph has a cycle
          items: { type: string }
        jobs:
          type: array
          description: what Run would submit; empty when the pipeline does not validate
          items: { $ref: "#/components/schemas/PipelineJobPreview" }
    PipelineRunResult:
      type: object
      description: The one group the whole pipeline became.
      required: [group_id, jobs]
      properties:
        group_id: { type: string }
        pipeline: { type: string }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/JobStatus" }
    PipelineListEntry:
      type: object
      required: [name, modified_at, size]
      properties:
        name: { type: string }
        modified_at: { type: number, description: unix seconds }
        size: { type: integer }
    PipelineKinds:
      type: object
      required: [port_types, kinds]
      properties:
        port_types:
          type: array
          items: { type: string }
        kinds:
          type: array
          items:
            type: object
            required: [kind, inputs, outputs, required]
            properties:
              kind: { type: string }
              inputs:
                type: array
                items: { type: string }
              outputs:
                type: array
                items: { type: string }
              required:
                type: array
                description: inputs that must be wired or satisfied by the node's own config
                items: { type: string }
...
    SimulationConfig:
      type: object
      additionalProperties: true
      x-tit-config: SimulationConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    Montage:
      type: object
      additionalProperties: true
      x-tit-config: Montage
      description: "injected from contracts/schema.json by dev/build_contract.py"
    FlexConfig:
      type: object
      additionalProperties: true
      x-tit-config: FlexConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    ExConfig:
      type: object
      additionalProperties: true
      x-tit-config: ExConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    MExConfig:
      type: object
      additionalProperties: true
      x-tit-config: MExConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    AnalyzerConfig:
      type: object
      additionalProperties: true
      x-tit-config: AnalyzerConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    PreprocessConfig:
      type: object
      additionalProperties: true
      x-tit-config: PreprocessConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    QSIPrepConfig:
      type: object
      additionalProperties: true
      x-tit-config: QSIPrepConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    QSIReconConfig:
      type: object
      additionalProperties: true
      x-tit-config: QSIReconConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    GroupComparisonConfig:
      type: object
      additionalProperties: true
      x-tit-config: GroupComparisonConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    CorrelationConfig:
      type: object
      additionalProperties: true
      x-tit-config: CorrelationConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    SourceConfig:
      type: object
      additionalProperties: true
      x-tit-config: SourceConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    LeadfieldConfig:
      type: object
      additionalProperties: true
      x-tit-config: LeadfieldConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    MontageConfig:
      type: object
      additionalProperties: true
      x-tit-config: MontageConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    VectorConfig:
      type: object
      additionalProperties: true
      x-tit-config: VectorConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    RegionConfig:
      type: object
      additionalProperties: true
      x-tit-config: RegionConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    NiftiAverageConfig:
      type: object
      additionalProperties: true
      x-tit-config: NiftiAverageConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    NilearnConfig:
      type: object
      additionalProperties: true
      x-tit-config: NilearnConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    PipelineConfig:
      description: >-
        Union of every config dataclass a job/validate/plan "config" body can carry. The real
        discriminant is the `kind`/path parameter alongside it, not a field on the object itself;
        this union exists so generated TS types cover every shape before dev/build_contract.py
        replaces each placeholder member with its generated schema.
      oneOf:
        - $ref: "#/components/schemas/SimulationConfig"
        - $ref: "#/components/schemas/Montage"
        - $ref: "#/components/schemas/FlexConfig"
        - $ref: "#/components/schemas/ExConfig"
        - $ref: "#/components/schemas/MExConfig"
        - $ref: "#/components/schemas/AnalyzerConfig"
        - $ref: "#/components/schemas/PreprocessConfig"
        - $ref: "#/components/schemas/QSIPrepConfig"
        - $ref: "#/components/schemas/QSIReconConfig"
        - $ref: "#/components/schemas/GroupComparisonConfig"
        - $ref: "#/components/schemas/CorrelationConfig"
        - $ref: "#/components/schemas/SourceConfig"
        - $ref: "#/components/schemas/LeadfieldConfig"
        - $ref: "#/components/schemas/MontageConfig"
        - $ref: "#/components/schemas/VectorConfig"
        - $ref: "#/components/schemas/RegionConfig"
        - $ref: "#/components/schemas/NiftiAverageConfig"
        - $ref: "#/components/schemas/NilearnConfig"