# tit/server/routes/schema.py
"""``/api/schema`` and ``/api/schema/{name}`` — serve ``contracts/schema.json`` (v1).

The file is generated by ``dev/build_schema.py`` (F1b) from the config
dataclasses; this module only reads and caches it by ``(path, mtime)`` so a
rebuild during development is picked up without a server restart.
"""

from __future__ import annotations

import json
from pathlib import Path
from typing import Any

from fastapi import APIRouter, HTTPException

import tit

router = APIRouter()

SCHEMA_PATH = Path(tit.__file__).resolve().parent.parent / "contracts" / "schema.json"

_cache: tuple[float, dict[str, Any]] | None = None


def _load_schema() -> dict[str, Any]:
    global _cache
    try:
        mtime = SCHEMA_PATH.stat().st_mtime
    except OSError as exc:
        raise HTTPException(
            status_code=503, detail=f"contracts/schema.json not found: {exc}"
        ) from exc
    if _cache is None or _cache[0] != mtime:
        with open(SCHEMA_PATH) as f:
            _cache = (mtime, json.load(f))
    return _cache[1]


@router.get("/api/schema", summary="The full generated schema.json")
def schema() -> dict[str, Any]:
    return _load_schema()


@router.get(
    "/api/schema/{name}",
    summary="One resolved $def from schema.json, by config class name",
)
def schema_by_name(name: str) -> dict[str, Any]:
    defs = _load_schema().get("$defs", {})
    if name not in defs:
        raise HTTPException(status_code=404, detail=f"Unknown config class: {name}")
    return defs[name]